+++ src/DripChip.Application/Features/Animals/Commands/Delete/DeleteAnimalCommandHandler.cs
@@ -1,6 +1,7 @@
 using DripChip.Application.Abstractions.Persistence;
 using DripChip.Application.Exceptions;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace DripChip.Application.Features.Animals.Commands.Delete;
 
@@ -14,8 +15,13 @@ public class DeleteAnimalCommandHandler : IRequestHandler<DeleteAnimalCommand>
     public async Task<Unit> Handle(DeleteAnimalCommand request, CancellationToken cancellationToken)
     {
         var animal =
-            await _context.Animals.FindAsync(request.Id)
+            await _context.Animals
+                .Include(animal => animal.VisitedLocations)
+                .FirstOrDefaultAsync(animal => animal.Id == request.Id, cancellationToken)
             ?? throw new NotFoundException();
+
+        if (animal.VisitedLocations.Any())
+            throw new ValidationException(nameof(request.Id), "The specified animal has moved away from its chipping location.");
 
         _context.Animals.Remove(animal);
         await _context.SaveChangesAsync(cancellationToken);
